feat: Add minimal Vale configuration, existing package files
This commit adds a new minimal Vale configuration, as well as some
package files from a prior iteration of a style guide for NGINX/Vale.
There is a lot of nuance to how Vale can be configured: the intent of
adding these files is so that we can begin to iterate on what to
override, add or disable alongside style guide work.